Extract prop settings
