+++ src/elements/ia-status-indicator/ia-status-indicator-story.ts
@@ -1,9 +1,9 @@
 import { html, LitElement } from 'lit';
 import { customElement, state } from 'lit/decorators.js';
 
-import { PropInputSettings } from '@demo/story-template';
+import type { PropInputSettings } from '@demo/story-components/story-prop-settings';
 import type { StyleInputSettings } from '@demo/story-components/story-styles-settings';
-import { IAStatusIndicator } from './ia-status-indicator';
+import type { IAStatusIndicator } from './ia-status-indicator';
 
 import './ia-status-indicator';
 import '@demo/story-template';
